move types & actions to separate files
