+++ front/src/applications/customget/components/ChartHelpers.js
@@ -1,7 +1,7 @@
 import * as d3 from 'd3';
 
 import { sec2time } from 'utils/timeManipulation';
-import { updateMustRedraw } from 'reducers/osrdsimulation';
+import { updateMustRedraw } from 'reducers/osrdsimulation/actions';
 
 export const sec2d3datetime = (time) => d3.timeParse('%H:%M:%S')(sec2time(time));
