+++ front/src/applications/customget/views/SpaceTimeChart.js
@@ -14,8 +14,8 @@ import {
   updateContextMenu,
   updateMustRedraw,
   updatePositionValues,
-  updateSimulation,
-} from 'reducers/osrdsimulation';
+  updateSimulation
+} from 'reducers/osrdsimulation/actions';
 import { useDispatch, useSelector } from 'react-redux';
 
 import { CgLoadbar } from 'react-icons/cg';
