+++ front/src/applications/osrd/components/Simulation/ChartModal.js
@@ -5,7 +5,7 @@ import { useDispatch, useSelector } from 'react-redux';
 
 import PropTypes from 'prop-types';
 import { changeTrain } from 'applications/osrd/components/TrainList/TrainListHelpers';
-import { updateMustRedraw } from 'reducers/osrdsimulation';
+import { updateMustRedraw } from 'reducers/osrdsimulation/actions';
 import { useTranslation } from 'react-i18next';
 
 function ChartModal(props) {
