+++ front/src/applications/osrd/components/Simulation/ContextMenu.js
@@ -11,7 +11,7 @@ import {
   updateMustRedraw,
   updateSelectedProjection,
   updateSelectedTrain,
-} from 'reducers/osrdsimulation';
+} from 'reducers/osrdsimulation/actions';
 import { useDispatch, useSelector } from 'react-redux';
 
 import CheckboxRadioSNCF from 'common/BootstrapSNCF/CheckboxRadioSNCF';
