+++ front/src/applications/osrd/components/Simulation/SignalSwitch/withOSRDData.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
 import { LIST_VALUES_SIGNAL_BASE } from 'applications/osrd/components/Simulation/consts';
-import { updateSignalBase } from 'reducers/osrdsimulation';
+import { updateSignalBase } from 'reducers/osrdsimulation/actions';
 import SignalSwitch from './SignalSwitch';
 
 /**
