+++ front/src/applications/osrd/components/Simulation/enableInteractivity.js
@@ -13,7 +13,7 @@ import {
   updateContextMenu,
   updateMustRedraw,
   updateTimePositionValues,
-} from 'reducers/osrdsimulation';
+} from 'reducers/osrdsimulation/actions';
 
 import { LIST_VALUES_NAME_SPACE_TIME } from 'applications/osrd/components/Simulation/consts';
 import drawGuideLines from 'applications/osrd/components/Simulation/drawGuideLines';
@@ -345,7 +345,7 @@ const enableInteractivity = (
   chart.svg // .selectAll('.zoomable')
     .on('mouseover', () => displayGuide(chart, 1))
     .on('mousemove', mousemove)
-    .on('wheel', (event) =>  {
+    .on('wheel', (event) => {
       if (event.ctrlKey || event.shiftKey) {
         event.preventDefault();
       }
